The site tracks which sections were open/closed.
The only annoyance that I experienced with the site was when I had the
schedule open, clicked a link to see some example code, and then
navigated back, the schedule section would be closed and I would have to
open it again before I could look at more example code. This keeps track
of which sections were open/closed and will maintain that state between
visits.